Financial analyst using agno and GPT-OSS coder/requirements.txt
mcp>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pydantic>=2.0.0
agno>=0.1.0
python-dotenv>=1.0.0